TextInput: Move PropTypes to DeprecatedPropTypes
